% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iguana.uncertainty.get.R
\docType{data}
\name{token}
\alias{token}
\title{Get uncertainty data from the Iguana API}
\format{An object of class \code{character} of length 1.}
\usage{
token
}
\arguments{
\item{token}{A \code{string}. The access code to use Iguana API}

\item{fonte}{PARAM_DESCRIPTION}

\item{datainicio}{PARAM_DESCRIPTION}

\item{datafim}{PARAM_DESCRIPTION}

\item{categoria}{PARAM_DESCRIPTION}

\item{output}{PARAM_DESCRIPTION, Default: c("error", "message", "data")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
 \code{\link[jsonlite]{fromJSON}}
 \code{\link[base]{rawToChar}}
}
\keyword{datasets}
