% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_encoding.R
\name{repair_encoding}
\alias{repair_encoding}
\title{FUNCTION_TITLE}
\usage{
repair_encoding(x, from = NULL)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{from}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
