% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iguana.get.R
\name{iguana.get}
\alias{iguana.get}
\title{Get data from the Iguana API}
\usage{
iguana.get(token, fonte, datainicio, datafim, categoria)
}
\arguments{
\item{token}{A \code{string}. The access code to use Iguana API}

\item{fonte}{PARAM_DESCRIPTION}

\item{datainicio}{PARAM_DESCRIPTION}

\item{datafim}{PARAM_DESCRIPTION}

\item{categoria}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function acts as a wrapper to the Iguana API, providing an easy access to the data through R.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{get}
\keyword{news}
